pets/Care-o-Pet/requirements.txt
adafruit_blinka_displayio==0.11.1
blinka_displayio_pygamedisplay==2.4.0
Pillow==11.1.0
pygame==2.6.1
